Decouple User IDs and Actor IDs
Prior to this change the ship controlled by a user is the one which as
the same actor ID as their network user ID. This change makes these
independent and a lookup is performed on the server to determine which
actor to update with a User's commands. Doing things this way reduces
brittleness (If a user accidentally selected an ID already sued by an
actor it would break).

To support this change, the following changes were also made:

* A world on the client now has a concept of local_user_id and
  local_ship_id.
* An actor is not initialized with it's owner, but rather deserializes
  it from it's update messages.
* When an actor is first created, it is done so along with an initial
  update containing it's full state.
* Actor type is a property of the base type, rather than a
  'free-floating' variable inside the actor's .lua file

The change in this patch and the smaller changes above will go on to
support respawning and other necessary gameplay concepts.